requirements.txt
awscli>=1.32.108
boto3>=1.34.108
click==8.1.8
detect-secrets==1.5.0
google-api-python-client>=2.88.0
matplotlib>=3.9.0
oci>=2.104.0
pandas>=2.2.0
pluginbase==1.0.1
psycopg2-binary>=2.9.9
pymongo>=4.6.1
pysnow<=0.7.17
python3-nmap>=1.6.0
tomli>=2.0.1
vt-py>=0.18.0
snowflake-connector-python>=3.12.1
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
gunicorn>=22.0
whitenoise>=6.5
python-dotenv>=1.0
celery>=5.4
requests>=2.32
python-jose[cryptography]>=3.3.0
crewai>=0.51,<1.0
crewai-tools>=0.13,<1.0
azure-identity
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
azure-mgmt-sql
azure-mgmt-resource
azure-mgmt-web
azure-mgmt-monitor
msrestazure
python-freshdesk
stripe
